+++ backend/src/main/java/com/volunteer/auth/dto/JwtUser.java
@@ -1,4 +1,4 @@
 package com.volunteer.auth.dto;
 
-public record JwtUser(Long id, String email) {
+public record JwtUser(Integer id, String email) {
 }
